Rename 'pre_post_process' folder to 'preprocess', aligned with namespace Fix 32-bit tests
