+++ ngraph/core/include/openvino/core/preprocess/input_tensor_info.hpp
