+++ ngraph/test/CMakeLists.txt
@@ -76,7 +76,7 @@ set(SRC
     pass_config.cpp
     pass_manager.cpp
     pattern.cpp
-    pre_post_process.cpp
+    preprocess.cpp
     provenance.cpp
     replace_node.cpp
     reshape_opt_kernel.cpp
